@@ -1,6 +1,6 @@
 {
   "name": "react-draggable-component",
-  "version": "4.4.0",
+  "version": "5.0.0",
   "description": "A wrapper React component that adds draggable functionality to an element",
   "main": "index.js",
   "author": "Mike Fey",
@@ -11,61 +11,65 @@
   },
   "scripts": {
     "test": "./node_modules/karma/bin/karma start",
-    "webpack-server": "webpack-dev-server --hot --progress --colors",
+    "webpack-server": "webpack-dev-server --hot --progress --colors --watch",
     "web-server": "http-server -p 3000 .",
     "start": "npm run webpack-server & npm run web-server",
-    "build": "NODE_ENV=production webpack -p --config webpack.production.config.js -d --display-reasons --display-chunks --progress"
+    "build": "NODE_ENV=production webpack --config webpack.production.config.js --display-reasons --display-chunks --progress"
   },
   "peerDependencies": {
-    "react": ">=15"
+    "react": ">=16"
   },
   "devDependencies": {
-    "autoprefixer": "^6.3.1",
-    "babel": "^6.3.26",
+    "autoprefixer": "^9.1.0",
+    "babel": "^6.23.0",
     "babel-cli": "^6.4.5",
     "babel-core": "^6.4.5",
-    "babel-loader": "^6.2.1",
+    "babel-loader": "^7.1.5",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
-    "css-loader": "^0.23.1",
+    "css-loader": "^1.0.0",
     "document-ready-promise": "^3.0.1",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^4.0.0",
-    "eslint-plugin-react": "^3.16.1",
+    "eslint": "^5.3.0",
+    "eslint-config-airbnb": "^17.0.0",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-react": "^7.10.0",
     "faucet": "0.0.1",
-    "file-loader": "^0.8.5",
-    "http-server": "^0.8.5",
-    "img-loader": "^1.2.2",
-    "karma": "^0.13.19",
-    "karma-chrome-launcher": "^0.2.2",
-    "karma-es6-shim": "^0.2.3",
+    "file-loader": "^1.1.11",
+    "http-server": "^0.11.1",
+    "img-loader": "^3.0.0",
+    "karma": "^2.0.5",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-es6-shim": "^1.0.0",
     "karma-phantomjs-launcher": "^1.0.0",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-tap": "^1.0.3",
-    "karma-webpack": "^1.7.0",
-    "node-sass": "^3.4.2",
+    "karma-tap": "^4.1.4",
+    "karma-webpack": "^3.0.0",
+    "node-sass": "^4.9.2",
     "object.assign": "^4.0.3",
     "phantomjs-polyfill-object-assign": "0.0.2",
     "phantomjs-prebuilt": "^2.1.3",
-    "postcss-loader": "^0.8.0",
-    "precss": "^1.4.0",
-    "react": "^15.3.2",
-    "react-addons-test-utils": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-hot-loader": "^1.3.0",
-    "react-unit": "^1.1.1",
-    "sass-loader": "^3.1.2",
-    "source-map-loader": "^0.1.5",
-    "style-loader": "^0.13.0",
-    "svgo-loader": "^1.1.0",
+    "precss": "^3.1.2",
+    "react": "^16.4.2",
+    "react-dom": "^16.4.2",
+    "react-hot-loader": "^4.3.4",
+    "react-unit": "^3.0.3",
+    "sass-loader": "^7.1.0",
+    "source-map-loader": "^0.2.3",
+    "style-loader": "^0.22.0",
+    "svgo-loader": "^2.1.0",
     "tape": "^4.4.0",
     "tape-jsx-equals": "^1.0.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.12.12",
-    "webpack-dev-server": "^1.14.1"
+    "uglifyjs-webpack-plugin": "^1.2.7",
+    "url-loader": "^1.0.1",
+    "webpack": "^4.16.5",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.5"
   },
   "dependencies": {
+    "postcss-loader": "^2.1.6",
     "prop-types": "^15.6.1"
   }
 }
