@@ -1,3 +1,4 @@
-var context = require.context('.', true, /-test\.js$/);
+const context = require.context('.', true, /-test\.js$/);
+
 context.keys().forEach(context);
-module.exports = context;
+export default context;
