@@ -1,46 +1,1 @@
-!function(e,t){"object"==typeof exports&&"object"==typeof module?module.exports=t(require("react")):"function"==typeof define&&define.amd?define(["react"],t):"object"==typeof exports?exports.ResponsiveImage=t(require("react")):e.ResponsiveImage=t(e.react)}(this,function(e){return function(e){function t(r){if(n[r])return n[r].exports;var o=n[r]={exports:{},id:r,loaded:!1};return e[r].call(o.exports,o,o.exports,t),o.loaded=!0,o.exports}var n={};return t.m=e,t.c=n,t.p="",t(0)}([/*!*******************************!*\
-  !*** ./assets/js/src/main.js ***!
-  \*******************************/
-function(e,t,n){"use strict";function r(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(t,"__esModule",{value:!0});var o=n(/*! ./components/Draggable.jsx */6),i=r(o);t.default=i.default},/*!******************************!*\
-  !*** ./~/process/browser.js ***!
-  \******************************/
-function(e,t){function n(){throw new Error("setTimeout has not been defined")}function r(){throw new Error("clearTimeout has not been defined")}function o(e){if(l===setTimeout)return setTimeout(e,0);if((l===n||!l)&&setTimeout)return l=setTimeout,setTimeout(e,0);try{return l(e,0)}catch(t){try{return l.call(null,e,0)}catch(t){return l.call(this,e,0)}}}function i(e){if(p===clearTimeout)return clearTimeout(e);if((p===r||!p)&&clearTimeout)return p=clearTimeout,clearTimeout(e);try{return p(e)}catch(t){try{return p.call(null,e)}catch(t){return p.call(this,e)}}}function a(){y&&d&&(y=!1,d.length?h=d.concat(h):v=-1,h.length&&s())}function s(){if(!y){var e=o(a);y=!0;for(var t=h.length;t;){for(d=h,h=[];++v<t;)d&&d[v].run();v=-1,t=h.length}d=null,y=!1,i(e)}}function u(e,t){this.fun=e,this.array=t}function c(){}var l,p,f=e.exports={};!function(){try{l="function"==typeof setTimeout?setTimeout:n}catch(e){l=n}try{p="function"==typeof clearTimeout?clearTimeout:r}catch(e){p=r}}();var d,h=[],y=!1,v=-1;f.nextTick=function(e){var t=new Array(arguments.length-1);if(arguments.length>1)for(var n=1;n<arguments.length;n++)t[n-1]=arguments[n];h.push(new u(e,t)),1!==h.length||y||o(s)},u.prototype.run=function(){this.fun.apply(null,this.array)},f.title="browser",f.browser=!0,f.env={},f.argv=[],f.version="",f.versions={},f.on=c,f.addListener=c,f.once=c,f.off=c,f.removeListener=c,f.removeAllListeners=c,f.emit=c,f.prependListener=c,f.prependOnceListener=c,f.listeners=function(e){return[]},f.binding=function(e){throw new Error("process.binding is not supported")},f.cwd=function(){return"/"},f.chdir=function(e){throw new Error("process.chdir is not supported")},f.umask=function(){return 0}},/*!*************************************!*\
-  !*** ./~/fbjs/lib/emptyFunction.js ***!
-  \*************************************/
-function(e,t){"use strict";function n(e){return function(){return e}}var r=function(){};r.thatReturns=n,r.thatReturnsFalse=n(!1),r.thatReturnsTrue=n(!0),r.thatReturnsNull=n(null),r.thatReturnsThis=function(){return this},r.thatReturnsArgument=function(e){return e},e.exports=r},/*!*********************************!*\
-  !*** ./~/fbjs/lib/invariant.js ***!
-  \*********************************/
-function(e,t,n){(function(t){"use strict";function n(e,t,n,o,i,a,s,u){if(r(t),!e){var c;if(void 0===t)c=new Error("Minified exception occurred; use the non-minified dev environment for the full error message and additional helpful warnings.");else{var l=[n,o,i,a,s,u],p=0;c=new Error(t.replace(/%s/g,function(){return l[p++]})),c.name="Invariant Violation"}throw c.framesToPop=1,c}}var r=function(e){};"production"!==t.env.NODE_ENV&&(r=function(e){if(void 0===e)throw new Error("invariant requires an error message argument")}),e.exports=n}).call(t,n(/*! ./../../process/browser.js */1))},/*!**************************************************!*\
-  !*** ./~/prop-types/lib/ReactPropTypesSecret.js ***!
-  \**************************************************/
-function(e,t){"use strict";var n="SECRET_DO_NOT_PASS_THIS_OR_YOU_WILL_BE_FIRED";e.exports=n},/*!*******************************!*\
-  !*** ./~/fbjs/lib/warning.js ***!
-  \*******************************/
-function(e,t,n){(function(t){"use strict";var r=n(/*! ./emptyFunction */2),o=r;if("production"!==t.env.NODE_ENV){var i=function(e){for(var t=arguments.length,n=Array(t>1?t-1:0),r=1;r<t;r++)n[r-1]=arguments[r];var o=0,i="Warning: "+e.replace(/%s/g,function(){return n[o++]});"undefined"!=typeof console&&console.error(i);try{throw new Error(i)}catch(e){}};o=function(e,t){if(void 0===t)throw new Error("`warning(condition, format, ...args)` requires a warning message argument");if(0!==t.indexOf("Failed Composite propType: ")&&!e){for(var n=arguments.length,r=Array(n>2?n-2:0),o=2;o<n;o++)r[o-2]=arguments[o];i.apply(void 0,[t].concat(r))}}}e.exports=o}).call(t,n(/*! ./../../process/browser.js */1))},/*!************************************************!*\
-  !*** ./assets/js/src/components/Draggable.jsx ***!
-  \************************************************/
-function(e,t,n){"use strict";function r(e){return e&&e.__esModule?e:{default:e}}function o(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function i(e,t){if(!e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!t||"object"!=typeof t&&"function"!=typeof t?e:t}function a(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function, not "+typeof t);e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),t&&(Object.setPrototypeOf?Object.setPrototypeOf(e,t):e.__proto__=t)}Object.defineProperty(t,"__esModule",{value:!0});var s=function(){function e(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}return function(t,n,r){return n&&e(t.prototype,n),r&&e(t,r),t}}(),u=n(/*! react */12),c=r(u),l=n(/*! prop-types */11),p=r(l),f=function(e){function t(e){o(this,t);var n=i(this,(t.__proto__||Object.getPrototypeOf(t)).call(this,e));return n.doDrag=!0,n.state={pos:n.props.position||{x:0,y:0},dragging:!1,elementStyle:{position:n.props.cssPosition||"absolute"},checkTimerStarted:!1,scrollLocked:!1,dragLocked:!1,dragStartInputPos:{x:0,y:0},dragStartElementPos:{x:0,y:0}},n.renderChild=n.renderChild.bind(n),n.setPosition=n.setPosition.bind(n),n.positionContent=n.positionContent.bind(n),n.onDragStart=n.onDragStart.bind(n),n.onDrag=n.onDrag.bind(n),n.onDragStop=n.onDragStop.bind(n),n.onDragLeave=n.onDragLeave.bind(n),n}return a(t,e),s(t,[{key:"componentWillMount",value:function(){this.props.position&&this.positionContent()}},{key:"componentDidMount",value:function(){document.addEventListener("mousemove",this.onDrag),document.addEventListener("touchmove",this.onDrag),document.addEventListener("mouseup",this.onDragStop),document.addEventListener("touchend",this.onDragStop)}},{key:"componentDidUpdate",value:function(e){var t=this.props.position;!t||e.position&&(!e.position||t.x===e.position.x&&t.y===e.position.y)||this.setPosition({x:t.x,y:t.y})}},{key:"componentWillUnmount",value:function(){document.removeEventListener("mousemove",this.onDrag),document.removeEventListener("touchmove",this.onDrag),document.removeEventListener("mouseup",this.onDragStop),document.removeEventListener("touchend",this.onDragStop)}},{key:"render",value:function(){var e=this.props.className||"",t="component-draggable "+e,n=this.renderChild();return c.default.createElement("div",{className:t,onTouchStart:this.onDragStart,onTouchCancel:this.onDragLeave,onMouseDown:this.onDragStart,onDragLeave:this.onDragLeave},n)}},{key:"renderChild",value:function(){var e=this,t=this;return c.default.Children.map(this.props.children,function(n){var r={},o={},i={};return n.props&&n.props.style&&(o=n.props.style),t.state&&t.state.elementStyle&&(i=t.state.elementStyle),r=Object.assign(o,i,e.props.style),c.default.cloneElement(n,{style:r,ref:"draggableChild"})})}},{key:"setPosition",value:function(e){this.setState({pos:e},this.positionContent)}},{key:"setScrollLock",value:function(e,t){var n=0,r=0;!this.props.touchScrollLock||"x"!==this.props.lock&&"y"!==this.props.lock||(r=t-this.state.pos.y,n=e-this.state.pos.x,r<0&&(r*=-1),n<0&&(n*=-1),"x"===this.props.lock?n>r?n>10&&!this.state.scrollLocked&&(this.doDrag=!0,this.setState({dragLocked:!0})):this.state.dragLocked||this.setState({scrollLocked:!0}):r>n?r>10&&!this.state.scrollLocked&&(this.doDrag=!0,this.setState({dragLocked:!0})):this.state.dragLocked||this.setState({scrollLocked:!0}))}},{key:"getNewPosition",value:function(e,t){var n={x:this.state.dragStartElementPos.x+(e-this.state.dragStartInputPos.x),y:this.state.dragStartElementPos.y+(t-this.state.dragStartInputPos.y)};if(this.props.bounds){var r=this.props.bounds;n.x<r.x1&&(n.x=r.x1),n.y<r.y1&&(n.y=r.y1),n.x>r.x2&&(n.x=r.x2),n.y>r.y2&&(n.y=r.y3)}return"x"===this.props.lock&&(n.y=this.state.dragStartElementPos.y),"y"===this.props.lock&&(n.x=this.state.dragStartElementPos.x),n}},{key:"positionContent",value:function(){var e=this,t=this.state.pos.x+"px",n=this.state.pos.y+"px",r="translate3d("+t+","+n+", 0px)",o={msTransform:r,WebkitTransform:r,MozTransform:r,transform:r},i=Object.assign(this.state.elementStyle,o);this.setState({elementStyle:i},function(){e.doDrag=!0})}},{key:"onDragStart",value:function(e){if(this.mouseDownOnElement=!0,this.props.disabled)this.mouseDownWhileDisabled=!0;else{var t=e.clientX,n=e.clientY;"mousedown"!==e.type&&(t=e.changedTouches[0].clientX,n=e.changedTouches[0].clientY),this.mouseDownWhileDisabled=!1,this.setState({scrollLocked:!1,checkTimerStarted:!0,dragStartInputPos:{x:t,y:n},dragStartElementPos:{x:this.state.pos.x,y:this.state.pos.y}}),this.props.dragStartCallback&&this.props.dragStartCallback(this.state.pos,e.target),this.props.preventDefaultEvents&&(e.stopPropagation(),e.preventDefault())}}},{key:"onDrag",value:function(e){var t=this;if(this.mouseDownOnElement&&!this.props.disabled){var n=e.clientX,r=e.clientY;if("mousemove"!==e.type&&(n=e.changedTouches[0].clientX,r=e.changedTouches[0].clientY),this.setScrollLock(n,r),this.doDrag&&!this.state.scrollLocked){this.doDrag=!1;var o=this.getNewPosition(n,r);this.setState({dragging:!0,pos:o},function(){t.positionContent()}),this.props.dragCallback&&this.props.dragCallback(o,event.target)}}this.props.preventDefaultEvents&&(e.stopPropagation(),e.preventDefault())}},{key:"onDragStop",value:function(e){!this.mouseDownOnElement||this.props.disabled||this.mouseDownWhileDisabled||(this.props.dragStopCallback&&!this.state.scrollLocked&&this.props.dragStopCallback({x:this.state.pos.x,y:this.state.pos.y},e.target),this.setState({dragging:!1,scrollLocked:!1,dragLocked:!1}),this.props.preventDefaultEvents&&(e.stopPropagation(),e.preventDefault())),this.mouseDownOnElement=!1}},{key:"onDragLeave",value:function(){this.state.dragging&&(this.setState({scrollLocked:!1,dragLocked:!1}),this.props.dragLeaveCallback&&this.props.dragLeaveCallback())}}]),t}(c.default.Component);f.propTypes={bounds:p.default.shape({x1:p.default.number.isRequired,y1:p.default.number.isRequired,x2:p.default.number.isRequired,y2:p.default.number.isRequired}),children:p.default.oneOfType([p.default.array,p.default.object]).isRequired,className:p.default.string,cssPosition:p.default.string,disabled:p.default.bool,dragCallback:p.default.func,dragLeaveCallback:p.default.func,dragStartCallback:p.default.func,dragStopCallback:p.default.func,lock:p.default.string,position:p.default.object,preventDefaultEvents:p.default.bool,style:p.default.object,touchScrollLock:p.default.bool},t.default=f},/*!**********************************!*\
-  !*** ./~/object-assign/index.js ***!
-  \**********************************/
-function(e,t){/*
-	object-assign
-	(c) Sindre Sorhus
-	@license MIT
-	*/
-"use strict";function n(e){if(null===e||void 0===e)throw new TypeError("Object.assign cannot be called with null or undefined");return Object(e)}function r(){try{if(!Object.assign)return!1;var e=new String("abc");if(e[5]="de","5"===Object.getOwnPropertyNames(e)[0])return!1;for(var t={},n=0;n<10;n++)t["_"+String.fromCharCode(n)]=n;var r=Object.getOwnPropertyNames(t).map(function(e){return t[e]});if("0123456789"!==r.join(""))return!1;var o={};return"abcdefghijklmnopqrst".split("").forEach(function(e){o[e]=e}),"abcdefghijklmnopqrst"===Object.keys(Object.assign({},o)).join("")}catch(e){return!1}}var o=Object.getOwnPropertySymbols,i=Object.prototype.hasOwnProperty,a=Object.prototype.propertyIsEnumerable;e.exports=r()?Object.assign:function(e,t){for(var r,s,u=n(e),c=1;c<arguments.length;c++){r=Object(arguments[c]);for(var l in r)i.call(r,l)&&(u[l]=r[l]);if(o){s=o(r);for(var p=0;p<s.length;p++)a.call(r,s[p])&&(u[s[p]]=r[s[p]])}}return u}},/*!****************************************!*\
-  !*** ./~/prop-types/checkPropTypes.js ***!
-  \****************************************/
-function(e,t,n){(function(t){"use strict";function r(e,n,r,u,c){if("production"!==t.env.NODE_ENV)for(var l in e)if(e.hasOwnProperty(l)){var p;try{o("function"==typeof e[l],"%s: %s type `%s` is invalid; it must be a function, usually from the `prop-types` package, but received `%s`.",u||"React class",r,l,typeof e[l]),p=e[l](n,l,u,r,null,a)}catch(e){p=e}if(i(!p||p instanceof Error,"%s: type specification of %s `%s` is invalid; the type checker function must return `null` or an `Error` but returned a %s. You may have forgotten to pass an argument to the type checker creator (arrayOf, instanceOf, objectOf, oneOf, oneOfType, and shape all require an argument).",u||"React class",r,l,typeof p),p instanceof Error&&!(p.message in s)){s[p.message]=!0;var f=c?c():"";i(!1,"Failed %s type: %s%s",r,p.message,null!=f?f:"")}}}if("production"!==t.env.NODE_ENV)var o=n(/*! fbjs/lib/invariant */3),i=n(/*! fbjs/lib/warning */5),a=n(/*! ./lib/ReactPropTypesSecret */4),s={};e.exports=r}).call(t,n(/*! ./../process/browser.js */1))},/*!**************************************************!*\
-  !*** ./~/prop-types/factoryWithThrowingShims.js ***!
-  \**************************************************/
-function(e,t,n){"use strict";var r=n(/*! fbjs/lib/emptyFunction */2),o=n(/*! fbjs/lib/invariant */3),i=n(/*! ./lib/ReactPropTypesSecret */4);e.exports=function(){function e(e,t,n,r,a,s){s!==i&&o(!1,"Calling PropTypes validators directly is not supported by the `prop-types` package. Use PropTypes.checkPropTypes() to call them. Read more at http://fb.me/use-check-prop-types")}function t(){return e}e.isRequired=e;var n={array:e,bool:e,func:e,number:e,object:e,string:e,symbol:e,any:e,arrayOf:t,element:e,instanceOf:t,node:e,objectOf:t,oneOf:t,oneOfType:t,shape:t,exact:t};return n.checkPropTypes=r,n.PropTypes=n,n}},/*!*************************************************!*\
-  !*** ./~/prop-types/factoryWithTypeCheckers.js ***!
-  \*************************************************/
-function(e,t,n){(function(t){"use strict";var r=n(/*! fbjs/lib/emptyFunction */2),o=n(/*! fbjs/lib/invariant */3),i=n(/*! fbjs/lib/warning */5),a=n(/*! object-assign */7),s=n(/*! ./lib/ReactPropTypesSecret */4),u=n(/*! ./checkPropTypes */8);e.exports=function(e,n){function c(e){var t=e&&(L&&e[L]||e[N]);if("function"==typeof t)return t}function l(e,t){return e===t?0!==e||1/e===1/t:e!==e&&t!==t}function p(e){this.message=e,this.stack=""}function f(e){function r(r,c,l,f,d,h,y){if(f=f||_,h=h||l,y!==s)if(n)o(!1,"Calling PropTypes validators directly is not supported by the `prop-types` package. Use `PropTypes.checkPropTypes()` to call them. Read more at http://fb.me/use-check-prop-types");else if("production"!==t.env.NODE_ENV&&"undefined"!=typeof console){var v=f+":"+l;!a[v]&&u<3&&(i(!1,"You are manually calling a React.PropTypes validation function for the `%s` prop on `%s`. This is deprecated and will throw in the standalone `prop-types` package. You may be seeing this warning due to a third-party PropTypes library. See https://fb.me/react-warning-dont-call-proptypes for details.",h,f),a[v]=!0,u++)}return null==c[l]?r?new p(null===c[l]?"The "+d+" `"+h+"` is marked as required "+("in `"+f+"`, but its value is `null`."):"The "+d+" `"+h+"` is marked as required in "+("`"+f+"`, but its value is `undefined`.")):null:e(c,l,f,d,h)}if("production"!==t.env.NODE_ENV)var a={},u=0;var c=r.bind(null,!1);return c.isRequired=r.bind(null,!0),c}function d(e){function t(t,n,r,o,i,a){var s=t[n],u=D(s);if(u!==e){var c=T(s);return new p("Invalid "+o+" `"+i+"` of type "+("`"+c+"` supplied to `"+r+"`, expected ")+("`"+e+"`."))}return null}return f(t)}function h(){return f(r.thatReturnsNull)}function y(e){function t(t,n,r,o,i){if("function"!=typeof e)return new p("Property `"+i+"` of component `"+r+"` has invalid PropType notation inside arrayOf.");var a=t[n];if(!Array.isArray(a)){var u=D(a);return new p("Invalid "+o+" `"+i+"` of type "+("`"+u+"` supplied to `"+r+"`, expected an array."))}for(var c=0;c<a.length;c++){var l=e(a,c,r,o,i+"["+c+"]",s);if(l instanceof Error)return l}return null}return f(t)}function v(){function t(t,n,r,o,i){var a=t[n];if(!e(a)){var s=D(a);return new p("Invalid "+o+" `"+i+"` of type "+("`"+s+"` supplied to `"+r+"`, expected a single ReactElement."))}return null}return f(t)}function g(e){function t(t,n,r,o,i){if(!(t[n]instanceof e)){var a=e.name||_,s=P(t[n]);return new p("Invalid "+o+" `"+i+"` of type "+("`"+s+"` supplied to `"+r+"`, expected ")+("instance of `"+a+"`."))}return null}return f(t)}function m(e){function n(t,n,r,o,i){for(var a=t[n],s=0;s<e.length;s++)if(l(a,e[s]))return null;var u=JSON.stringify(e);return new p("Invalid "+o+" `"+i+"` of value `"+a+"` "+("supplied to `"+r+"`, expected one of "+u+"."))}return Array.isArray(e)?f(n):("production"!==t.env.NODE_ENV?i(!1,"Invalid argument supplied to oneOf, expected an instance of array."):void 0,r.thatReturnsNull)}function b(e){function t(t,n,r,o,i){if("function"!=typeof e)return new p("Property `"+i+"` of component `"+r+"` has invalid PropType notation inside objectOf.");var a=t[n],u=D(a);if("object"!==u)return new p("Invalid "+o+" `"+i+"` of type "+("`"+u+"` supplied to `"+r+"`, expected an object."));for(var c in a)if(a.hasOwnProperty(c)){var l=e(a,c,r,o,i+"."+c,s);if(l instanceof Error)return l}return null}return f(t)}function k(e){function n(t,n,r,o,i){for(var a=0;a<e.length;a++){var u=e[a];if(null==u(t,n,r,o,i,s))return null}return new p("Invalid "+o+" `"+i+"` supplied to "+("`"+r+"`."))}if(!Array.isArray(e))return"production"!==t.env.NODE_ENV?i(!1,"Invalid argument supplied to oneOfType, expected an instance of array."):void 0,r.thatReturnsNull;for(var o=0;o<e.length;o++){var a=e[o];if("function"!=typeof a)return i(!1,"Invalid argument supplied to oneOfType. Expected an array of check functions, but received %s at index %s.",j(a),o),r.thatReturnsNull}return f(n)}function x(){function e(e,t,n,r,o){return E(e[t])?null:new p("Invalid "+r+" `"+o+"` supplied to "+("`"+n+"`, expected a ReactNode."))}return f(e)}function w(e){function t(t,n,r,o,i){var a=t[n],u=D(a);if("object"!==u)return new p("Invalid "+o+" `"+i+"` of type `"+u+"` "+("supplied to `"+r+"`, expected `object`."));for(var c in e){var l=e[c];if(l){var f=l(a,c,r,o,i+"."+c,s);if(f)return f}}return null}return f(t)}function O(e){function t(t,n,r,o,i){var u=t[n],c=D(u);if("object"!==c)return new p("Invalid "+o+" `"+i+"` of type `"+c+"` "+("supplied to `"+r+"`, expected `object`."));var l=a({},t[n],e);for(var f in l){var d=e[f];if(!d)return new p("Invalid "+o+" `"+i+"` key `"+f+"` supplied to `"+r+"`.\nBad object: "+JSON.stringify(t[n],null,"  ")+"\nValid keys: "+JSON.stringify(Object.keys(e),null,"  "));var h=d(u,f,r,o,i+"."+f,s);if(h)return h}return null}return f(t)}function E(t){switch(typeof t){case"number":case"string":case"undefined":return!0;case"boolean":return!t;case"object":if(Array.isArray(t))return t.every(E);if(null===t||e(t))return!0;var n=c(t);if(!n)return!1;var r,o=n.call(t);if(n!==t.entries){for(;!(r=o.next()).done;)if(!E(r.value))return!1}else for(;!(r=o.next()).done;){var i=r.value;if(i&&!E(i[1]))return!1}return!0;default:return!1}}function S(e,t){return"symbol"===e||("Symbol"===t["@@toStringTag"]||"function"==typeof Symbol&&t instanceof Symbol)}function D(e){var t=typeof e;return Array.isArray(e)?"array":e instanceof RegExp?"object":S(t,e)?"symbol":t}function T(e){if("undefined"==typeof e||null===e)return""+e;var t=D(e);if("object"===t){if(e instanceof Date)return"date";if(e instanceof RegExp)return"regexp"}return t}function j(e){var t=T(e);switch(t){case"array":case"object":return"an "+t;case"boolean":case"date":case"regexp":return"a "+t;default:return t}}function P(e){return e.constructor&&e.constructor.name?e.constructor.name:_}var L="function"==typeof Symbol&&Symbol.iterator,N="@@iterator",_="<<anonymous>>",C={array:d("array"),bool:d("boolean"),func:d("function"),number:d("number"),object:d("object"),string:d("string"),symbol:d("symbol"),any:h(),arrayOf:y,element:v(),instanceOf:g,node:x(),objectOf:b,oneOf:m,oneOfType:k,shape:w,exact:O};return p.prototype=Error.prototype,C.checkPropTypes=u,C.PropTypes=C,C}}).call(t,n(/*! ./../process/browser.js */1))},/*!*******************************!*\
-  !*** ./~/prop-types/index.js ***!
-  \*******************************/
-function(e,t,n){(function(t){if("production"!==t.env.NODE_ENV){var r="function"==typeof Symbol&&Symbol.for&&Symbol.for("react.element")||60103,o=function(e){return"object"==typeof e&&null!==e&&e.$$typeof===r},i=!0;e.exports=n(/*! ./factoryWithTypeCheckers */10)(o,i)}else e.exports=n(/*! ./factoryWithThrowingShims */9)()}).call(t,n(/*! ./../process/browser.js */1))},/*!************************!*\
-  !*** external "react" ***!
-  \************************/
-function(t,n){t.exports=e}])});
-//# sourceMappingURL=react-draggable.build.js.map
+!function(n,e){"object"==typeof exports&&"object"==typeof module?module.exports=e(require("react"),require("prop-types")):"function"==typeof define&&define.amd?define(["react","prop-types"],e):"object"==typeof exports?exports.ReactDraggable=e(require("react"),require("prop-types")):n.ReactDraggable=e(n.react,n["prop-types"])}(window,function(__WEBPACK_EXTERNAL_MODULE__2__,__WEBPACK_EXTERNAL_MODULE__3__){return function(n){var e={};function t(o){if(e[o])return e[o].exports;var r=e[o]={i:o,l:!1,exports:{}};return n[o].call(r.exports,r,r.exports,t),r.l=!0,r.exports}return t.m=n,t.c=e,t.d=function(n,e,o){t.o(n,e)||Object.defineProperty(n,e,{enumerable:!0,get:o})},t.r=function(n){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(n,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(n,"__esModule",{value:!0})},t.t=function(n,e){if(1&e&&(n=t(n)),8&e)return n;if(4&e&&"object"==typeof n&&n&&n.__esModule)return n;var o=Object.create(null);if(t.r(o),Object.defineProperty(o,"default",{enumerable:!0,value:n}),2&e&&"string"!=typeof n)for(var r in n)t.d(o,r,function(e){return n[e]}.bind(null,r));return o},t.n=function(n){var e=n&&n.__esModule?function(){return n.default}:function(){return n};return t.d(e,"a",e),e},t.o=function(n,e){return Object.prototype.hasOwnProperty.call(n,e)},t.p="",t(t.s=0)}([function(module,exports,__webpack_require__){"use strict";eval('\n\nObject.defineProperty(exports, "__esModule", {\n  value: true\n});\n\nvar _Draggable = __webpack_require__(1);\n\nvar _Draggable2 = _interopRequireDefault(_Draggable);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.default = _Draggable2.default;\n\n//# sourceURL=webpack://ReactDraggable/./assets/js/src/main.js?')},function(module,exports,__webpack_require__){"use strict";eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(2);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _propTypes = __webpack_require__(3);\n\nvar _propTypes2 = _interopRequireDefault(_propTypes);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Draggable = function (_React$Component) {\n  _inherits(Draggable, _React$Component);\n\n  function Draggable(props) {\n    _classCallCheck(this, Draggable);\n\n    // if the component drag should take place\n    var _this = _possibleConstructorReturn(this, (Draggable.__proto__ || Object.getPrototypeOf(Draggable)).call(this, props));\n\n    _this.doDrag = true;\n\n    var _this$props = _this.props,\n        cssPosition = _this$props.cssPosition,\n        position = _this$props.position;\n\n    // initial state object\n\n    _this.state = {\n      pos: position || { x: 0, y: 0 },\n      dragging: false,\n      elementStyle: {\n        position: cssPosition || 'absolute'\n      },\n      scrollLocked: false,\n      dragLocked: false,\n      dragStartInputPos: {\n        x: 0,\n        y: 0\n      },\n      dragStartElementPos: {\n        x: 0,\n        y: 0\n      }\n    };\n\n    // bind functions to component\n    _this.renderChild = _this.renderChild.bind(_this);\n    _this.setPosition = _this.setPosition.bind(_this);\n    _this.positionContent = _this.positionContent.bind(_this);\n    _this.onDragStart = _this.onDragStart.bind(_this);\n    _this.onDrag = _this.onDrag.bind(_this);\n    _this.onDragStop = _this.onDragStop.bind(_this);\n    _this.onDragLeave = _this.onDragLeave.bind(_this);\n    return _this;\n  }\n\n  /**\n   * Called before component mounts to DOM\n   * @returns {undefined} undefined\n   */\n\n\n  _createClass(Draggable, [{\n    key: 'componentWillMount',\n    value: function componentWillMount() {\n      var position = this.props.position;\n\n\n      if (position) {\n        this.positionContent();\n      }\n    }\n\n    /**\n     * Called after component mounts to DOM\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'componentDidMount',\n    value: function componentDidMount() {\n      document.addEventListener('mousemove', this.onDrag);\n      document.addEventListener('touchmove', this.onDrag);\n      document.addEventListener('mouseup', this.onDragStop);\n      document.addEventListener('touchend', this.onDragStop);\n    }\n\n    /**\n     * Called when component updates\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'componentDidUpdate',\n    value: function componentDidUpdate(prevProps) {\n      var position = this.props.position;\n\n\n      if (position && (!prevProps.position || prevProps.position && (position.x !== prevProps.position.x || position.y !== prevProps.position.y))) {\n        this.setPosition({ x: position.x, y: position.y });\n      }\n    }\n\n    /**\n     * Called before component is removed from the DOM\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'componentWillUnmount',\n    value: function componentWillUnmount() {\n      document.removeEventListener('mousemove', this.onDrag);\n      document.removeEventListener('touchmove', this.onDrag);\n      document.removeEventListener('mouseup', this.onDragStop);\n      document.removeEventListener('touchend', this.onDragStop);\n    }\n\n    /**\n     * Renders component\n     * @returns {Object} React element\n     */\n\n  }, {\n    key: 'render',\n    value: function render() {\n      var className = this.props.className;\n\n      var classNameProp = className || '';\n      var newClassName = 'component-draggable ' + classNameProp;\n      var children = this.renderChild();\n\n      return _react2.default.createElement(\n        'div',\n        {\n          className: newClassName,\n          onTouchStart: this.onDragStart,\n          onTouchCancel: this.onDragLeave,\n          onMouseDown: this.onDragStart,\n          onDragLeave: this.onDragLeave,\n          role: 'draggableelement'\n        },\n        children\n      );\n    }\n\n    /**\n     * Renders the child element\n     * @returns {Object} React element\n     */\n\n  }, {\n    key: 'renderChild',\n    value: function renderChild() {\n      var inst = this;\n      var _props = this.props,\n          children = _props.children,\n          style = _props.style;\n      var elementStyle = inst.state.elementStyle;\n\n\n      return _react2.default.Children.map(children, function (child) {\n        var newStyle = {};\n        var childStyle = {};\n        var newElementStyle = {};\n\n        if (child.props && child.props.style) {\n          childStyle = child.props.style;\n        }\n\n        if (inst.state && elementStyle) {\n          newElementStyle = elementStyle;\n        }\n\n        newStyle = Object.assign(childStyle, newElementStyle, style);\n\n        return _react2.default.cloneElement(child, {\n          style: newStyle,\n          ref: 'draggableChild'\n        });\n      });\n    }\n\n    /**\n     * Positions the draggable element\n     * @param {Object} pos - An object with x and y position values\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'setPosition',\n    value: function setPosition(pos) {\n      this.setState({\n        pos: pos\n      }, this.positionContent);\n    }\n\n    /**\n     * If the 'touchScrollLock' prop is set to true, and dragging is locked\n     * to the 'x' or 'y' axis, and the user is dragging\n     * (or scrolling on mobile) on the opposite axis, stop any dragging.\n     * This is so the content doesn't move when a user is scrolling on a\n     * mobile device.\n     * @param {Number} pageX - The x position of mouse or touch event\n     * @param {Number} pageY - The y position of mouse or touch event\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'setScrollLock',\n    value: function setScrollLock(pageX, pageY) {\n      var _props2 = this.props,\n          touchScrollLock = _props2.touchScrollLock,\n          lock = _props2.lock;\n      var _state = this.state,\n          dragLocked = _state.dragLocked,\n          pos = _state.pos,\n          scrollLocked = _state.scrollLocked;\n\n\n      var difX = 0;\n      var difY = 0;\n\n      if (touchScrollLock && (lock === 'x' || lock === 'y')) {\n        difY = pageY - pos.y;\n        difX = pageX - pos.x;\n\n        if (difY < 0) {\n          difY *= -1;\n        }\n\n        if (difX < 0) {\n          difX *= -1;\n        }\n\n        if (lock === 'x') {\n          // if the content is locked to dragging horizontally\n          if (difX > difY) {\n            // if the user is intending to swipe horizontally\n            if (difX > 10 && !scrollLocked) {\n              this.doDrag = true;\n              this.setState({ dragLocked: true });\n            }\n          } else if (!dragLocked) {\n            // if the user is intending to swipe vertically\n            this.setState({ scrollLocked: true });\n          }\n        } else if (difY > difX) {\n          // if the content is locked to dragging vertically\n          // if the user is intending to swipe vertically\n          if (difY > 10 && !scrollLocked) {\n            this.doDrag = true;\n            this.setState({ dragLocked: true });\n          }\n        } else if (!dragLocked) {\n          this.setState({ scrollLocked: true });\n        }\n      }\n    }\n\n    /**\n     * Gets the updated position values\n     * @param {Number} pageX - The x position of mouse or touch event\n     * @param {Number} pageY - The y position of mouse or touch event\n     * @returns {Object} The x and y coordinates to move the content to\n     */\n\n  }, {\n    key: 'getNewPosition',\n    value: function getNewPosition(pageX, pageY) {\n      var _props3 = this.props,\n          bounds = _props3.bounds,\n          lock = _props3.lock;\n      var _state2 = this.state,\n          dragStartElementPos = _state2.dragStartElementPos,\n          dragStartInputPos = _state2.dragStartInputPos;\n\n      var newPos = {\n        x: dragStartElementPos.x + (pageX - dragStartInputPos.x),\n        y: dragStartElementPos.y + (pageY - dragStartInputPos.y)\n      };\n\n      if (bounds) {\n        // if the bounds were supplied, don't let the content move outside\n        // of them\n        var newBounds = bounds;\n\n        if (newPos.x < newBounds.x1) {\n          newPos.x = newBounds.x1;\n        }\n\n        if (newPos.y < newBounds.y1) {\n          newPos.y = newBounds.y1;\n        }\n\n        if (newPos.x > newBounds.x2) {\n          newPos.x = newBounds.x2;\n        }\n\n        if (newPos.y > newBounds.y2) {\n          newPos.y = newBounds.y3;\n        }\n      }\n\n      if (lock === 'x') {\n        newPos.y = dragStartElementPos.y;\n      }\n\n      if (lock === 'y') {\n        newPos.x = dragStartElementPos.x;\n      }\n\n      return newPos;\n    }\n\n    /**\n     * Positions the draggable element\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'positionContent',\n    value: function positionContent() {\n      var _this2 = this;\n\n      var _state3 = this.state,\n          elementStyle = _state3.elementStyle,\n          pos = _state3.pos;\n\n      var xPos = pos.x + 'px';\n      var yPos = pos.y + 'px';\n      var positionTransformString = 'translate3d(' + xPos + ',' + yPos + ', 0px)';\n      var transformStyle = {\n        msTransform: positionTransformString,\n        WebkitTransform: positionTransformString,\n        MozTransform: positionTransformString,\n        transform: positionTransformString\n      };\n      var newStyle = Object.assign(elementStyle, transformStyle);\n\n      this.setState({\n        elementStyle: newStyle\n      }, function () {\n        // allow the component to be dragged again because all values have been\n        // updated through the React lifecycle\n        _this2.doDrag = true;\n      });\n    }\n\n    /**\n     * Called when the user's mouse is pressed\n     * @param {Event} e A mousedown event\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'onDragStart',\n    value: function onDragStart(e) {\n      var _props4 = this.props,\n          disabled = _props4.disabled,\n          dragStartCallback = _props4.dragStartCallback,\n          preventDefaultEvents = _props4.preventDefaultEvents;\n      var pos = this.state.pos;\n\n      this.mouseDownOnElement = true;\n\n      if (!disabled) {\n        var pageX = e.clientX;\n        var pageY = e.clientY;\n\n        if (e.type !== 'mousedown') {\n          pageX = e.changedTouches[0].clientX;\n          pageY = e.changedTouches[0].clientY;\n        }\n\n        this.mouseDownWhileDisabled = false;\n\n        this.setState({\n          scrollLocked: false,\n          dragStartInputPos: {\n            x: pageX,\n            y: pageY\n          },\n          dragStartElementPos: {\n            x: pos.x,\n            y: pos.y\n          }\n        });\n\n        if (dragStartCallback) {\n          dragStartCallback(pos, e.target);\n        }\n\n        if (preventDefaultEvents) {\n          e.stopPropagation();\n          e.preventDefault();\n        }\n      } else {\n        this.mouseDownWhileDisabled = true;\n      }\n    }\n\n    /**\n     * Called when the user's mouse is moved\n     * @param {Event} e A mousemove event\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'onDrag',\n    value: function onDrag(e) {\n      var _this3 = this;\n\n      var _props5 = this.props,\n          disabled = _props5.disabled,\n          dragCallback = _props5.dragCallback,\n          preventDefaultEvents = _props5.preventDefaultEvents;\n      var scrollLocked = this.state.scrollLocked;\n\n\n      if (this.mouseDownOnElement && !disabled) {\n        var pageX = e.clientX;\n        var pageY = e.clientY;\n\n        if (e.type !== 'mousemove') {\n          pageX = e.changedTouches[0].clientX;\n          pageY = e.changedTouches[0].clientY;\n        }\n\n        this.setScrollLock(pageX, pageY);\n\n        if (this.doDrag && !scrollLocked) {\n          this.doDrag = false;\n\n          var newPos = this.getNewPosition(pageX, pageY);\n\n          this.setState({\n            dragging: true,\n            pos: newPos\n          }, function () {\n            _this3.positionContent();\n          });\n\n          if (dragCallback) {\n            dragCallback(newPos, e.target);\n          }\n        }\n      }\n\n      if (preventDefaultEvents) {\n        e.stopPropagation();\n        e.preventDefault();\n      }\n    }\n\n    /**\n     * Called when the user's mouse is released\n     * @param {Event} e A mouseup event\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'onDragStop',\n    value: function onDragStop(e) {\n      var _props6 = this.props,\n          disabled = _props6.disabled,\n          dragStopCallback = _props6.dragStopCallback,\n          preventDefaultEvents = _props6.preventDefaultEvents;\n      var _state4 = this.state,\n          pos = _state4.pos,\n          scrollLocked = _state4.scrollLocked;\n\n\n      if (this.mouseDownOnElement && !disabled && !this.mouseDownWhileDisabled) {\n        if (dragStopCallback && !scrollLocked) {\n          dragStopCallback({\n            x: pos.x,\n            y: pos.y\n          }, e.target);\n        }\n\n        this.setState({\n          dragging: false,\n          scrollLocked: false,\n          dragLocked: false\n        });\n\n        if (preventDefaultEvents) {\n          e.stopPropagation();\n          e.preventDefault();\n        }\n      }\n\n      this.mouseDownOnElement = false;\n    }\n\n    /**\n     * Called when the user is dragging but releases the mouse\n     * outside of the draggable component\n     * @returns {undefined} undefined\n     */\n\n  }, {\n    key: 'onDragLeave',\n    value: function onDragLeave() {\n      var dragLeaveCallback = this.props.dragLeaveCallback;\n      var dragging = this.state.dragging;\n\n\n      if (dragging) {\n        this.setState({\n          scrollLocked: false,\n          dragLocked: false\n        });\n\n        if (dragLeaveCallback) {\n          dragLeaveCallback();\n        }\n      }\n    }\n  }]);\n\n  return Draggable;\n}(_react2.default.Component);\n\nDraggable.defaultProps = {\n  bounds: null,\n  className: null,\n  cssPosition: null,\n  disabled: false,\n  dragCallback: null,\n  dragLeaveCallback: null,\n  dragStartCallback: null,\n  dragStopCallback: null,\n  lock: null,\n  position: null,\n  preventDefaultEvents: false,\n  style: null,\n  touchScrollLock: false\n};\n\n/**\n * Expected propTypes\n * @prop {Object} bounds - An array of coordinates, forming a square, that the\n * user cannot drag the component outside of\n * @prop {Object} children - Child React elements\n * @prop {String} className - A string of additional classnames to add\n * to the element\n * @prop {String} cssPosition - The css positioning for for the element\n * (i.e. 'absolute' or 'fixed', defaults to 'absolute')\n * @prop {Boolean} disabled - If the component is disabled\n * @prop {Function} dragCallback - A callback function while the user is\n * dragging\n * @prop {Function} dragStartCallback - A callback function for when the user\n * stops dragging\n * @prop {Function} dragStopCallback - A callback function for when the user\n * stops dragging\n * @prop {Function} dragLeaveCallback - A callback function for when the user\n * is dragging and the mouse/touch leaves the draggable component\n * @prop {String} lock - An axis to lock element to when dragging, either\n * 'x' or 'y'\n * @prop {Boolean} preventDefaultEvents - Whether to prevent default\n * mouse/touch events\n * @prop {Object} style - A style object\n * @prop {String} touchScrollLock - If set to true, prevents the content from\n * being dragged if the user is scrolling in the opposite direction on a touch\n * device\n */\nDraggable.propTypes = {\n  bounds: _propTypes2.default.shape({\n    x1: _propTypes2.default.number.isRequired,\n    y1: _propTypes2.default.number.isRequired,\n    x2: _propTypes2.default.number.isRequired,\n    y2: _propTypes2.default.number.isRequired\n  }),\n  children: _propTypes2.default.oneOfType([_propTypes2.default.array, _propTypes2.default.object]).isRequired,\n  className: _propTypes2.default.string,\n  cssPosition: _propTypes2.default.string,\n  disabled: _propTypes2.default.bool,\n  dragCallback: _propTypes2.default.func,\n  dragLeaveCallback: _propTypes2.default.func,\n  dragStartCallback: _propTypes2.default.func,\n  dragStopCallback: _propTypes2.default.func,\n  lock: _propTypes2.default.string,\n  position: _propTypes2.default.object,\n  preventDefaultEvents: _propTypes2.default.bool,\n  style: _propTypes2.default.object,\n  touchScrollLock: _propTypes2.default.bool\n};\n\nexports.default = Draggable;\n\n//# sourceURL=webpack://ReactDraggable/./assets/js/src/components/Draggable.jsx?")},function(module,exports){eval("module.exports = __WEBPACK_EXTERNAL_MODULE__2__;\n\n//# sourceURL=webpack://ReactDraggable/external_%22react%22?")},function(module,exports){eval("module.exports = __WEBPACK_EXTERNAL_MODULE__3__;\n\n//# sourceURL=webpack://ReactDraggable/external_%22prop-types%22?")}])});
