@@ -1,49 +1,36 @@
+const path = require('path');
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
+
 module.exports = {
-  entry: './assets/js/src/main.js',
+  mode: 'production',
+  entry: path.join(__dirname, '/assets/js/src/main.js'),
   output: {
-    path: 'assets/js/lib/',
+    path: path.join(__dirname, 'assets/js/lib/'),
     filename: 'react-draggable.build.js',
     libraryTarget: 'umd',
-    library: 'ResponsiveImage'
+    library: 'ReactDraggable',
   },
   externals: [{
-    'react' : 'react',
-    'react-dom' : 'react-dom'
+    react: 'react',
+    'react-dom': 'react-dom',
+    'prop-types': 'prop-types',
   }],
   devtool: 'eval',
   module: {
-    preLoaders: [
+    rules: [
       {
         test: /\.jsx?$/,
         exclude: /(node_modules|bower_components)/,
-        loader: 'source-map'
-      }
+        loaders: [
+          'babel-loader?presets[]=stage-0,presets[]=react,presets[]=es2015',
+        ],
+      },
     ],
-    loaders: [
-      {
-        test: /\.scss$/,
-        include: /src/,
-        loaders: [
-          'style',
-          'css',
-          'postcss-loader',
-          'sass?outputStyle=expanded'
-        ]
-      },
-      {
-        test: /\.(jpe?g|png|gif|svg)$/i,
-        loaders: [
-          'url?limit=8192',
-          'img'
-        ]
-      },
-      {
-        test: /\.jsx?$/,
-        exclude: /(node_modules|bower_components)/,
-        loaders: [
-          'babel?presets[]=stage-0,presets[]=react,presets[]=es2015'
-        ]
-      }
-    ]
-  }
+  },
+  optimization: {
+    minimize: true,
+    minimizer: [
+      new UglifyJsPlugin(),
+    ],
+  },
 };
